services/elpyfi-core/elpyfi-engine/requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
alpaca-py==0.13.3
python-dotenv==1.0.0